requirements.txt
Flask==2.1.3
Flask-Migrate==3.1.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
pip==22.1.2
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dotenv==0.20.0
requests==2.28.1
SQLAlchemy==1.4.39